+++ src/main/java/com/researchspace/api/clientmodel/AbstractExportPost.java
@@ -1,5 +1,6 @@
 package com.researchspace.api.clientmodel;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,11 +13,13 @@ public class AbstractExportPost {
 	private ExportScope scope = ExportScope.USER;
 	private ExportFormat format =  ExportFormat.HTML;
 	
-	public String scope () {
+	@JsonIgnore
+	public String getScopeAsString() {
 		return scope.name().toLowerCase();
 	}
-	
-	public String format () {
+
+	@JsonIgnore
+	public String getFormatAsString() {
 		return format.name().toLowerCase();
 	}