+++ src/main/java/com/researchspace/api/clientmodel/ActivitySearchResult.java
@@ -4,16 +4,20 @@ package com.researchspace.api.clientmodel;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import lombok.ToString;
-import lombok.Value;
 
 /**
  * ActivitySearchResults
  */
-@Value
-@EqualsAndHashCode(callSuper=true)
-@ToString(callSuper=true)
+@Data
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ActivitySearchResult extends PaginatedResultList {
 
     private List<Activity> activities = new ArrayList<>();