+++ src/main/java/com/researchspace/api/clientmodel/FieldPut.java
@@ -1,29 +1,32 @@
 package com.researchspace.api.clientmodel;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 
-@Data
-@EqualsAndHashCode(callSuper=false)
-@AllArgsConstructor
-@NoArgsConstructor
 /**
- * Extends FieldPost with a field Id property to specify the Field whose content is to be updated. 
- * @author rspace
- *
+ * Extends FieldPost with a field Id property to specify the Field whose content is to be updated.
  */
+@Data
+@EqualsAndHashCode(callSuper=true)
+@NoArgsConstructor
 public class FieldPut extends FieldPost {
-	/**
-	 * Optional content, can be empty.
-	 */
-	private String content="";
 	/**
 	 * Cannot be null
 	 */
 	@NonNull
 	private Long id;
+
+	/** Constructor for updating a specific field by ID only (content left as default). */
+	public FieldPut(Long id) {
+		this.id = id;
+	}
+
+	/** Constructor for updating the content of a specific field by ID. */
+	public FieldPut(String content, Long id) {
+		super(content);
+		this.id = id;
+	}
 
 }