+++ src/main/java/com/researchspace/api/clientmodel/GroupSearchResult.java
@@ -1,24 +1,22 @@
 package com.researchspace.api.clientmodel;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /**
- * FileSearchResult
+ * Paginated list of groups returned by the sysadmin group-listing endpoint.
  */
 @Data
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class FileSearchResult extends PaginatedResultList {
-
-    private List<ApiFile> files = new ArrayList<>();
-
+public class GroupSearchResult extends PaginatedResultList {
+    private List<GroupInfo> groups = new ArrayList<>();
 }