+++ src/main/java/com/researchspace/api/clientmodel/IdentifiableNameable.java
@@ -1,15 +1,15 @@
 package com.researchspace.api.clientmodel;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
-
-import java.util.List;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @SuperBuilder
+@JsonIgnoreProperties(ignoreUnknown = true)
 public abstract class IdentifiableNameable extends Linkable {
     private Long id;
     private String globalId;