+++ src/main/java/com/researchspace/api/clientmodel/UserGroupInfo.java
@@ -1,8 +1,12 @@
 package com.researchspace.api.clientmodel;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class UserGroupInfo {
 	
 	private Long id;